+++ projects/rocprofiler-compute/src/lib/rocprofiler_compute_tool/dispatch_callback.h
@@ -1,7 +1,108 @@
 // Copyright (c) Advanced Micro Devices, Inc.
 // SPDX-License-Identifier:  MIT
+#pragma once
+
+#include "helper.hpp"
+
 #include <rocprofiler-sdk/registration.h>
 #include <rocprofiler-sdk/rocprofiler.h>
+
+#include <cstdint>
+#include <map>
+#include <memory>
+#include <mutex>
+#include <set>
+#include <shared_mutex>
+#include <sstream>
+#include <string>
+#include <unordered_map>
+#include <vector>
+
+#define ROCPROFILER_CALL(result, msg)                                                                  \
+    {                                                                                                  \
+        rocprofiler_status_t CHECKSTATUS = result;                                                     \
+        if (CHECKSTATUS != ROCPROFILER_STATUS_SUCCESS)                                                 \
+        {                                                                                              \
+            std::string status_msg = rocprofiler_get_status_string(CHECKSTATUS);                       \
+            std::cerr << "[" #result "][" << __FILE__ << ":" << __LINE__ << "] " << msg                \
+                      << " failed with error code " << CHECKSTATUS << ": " << status_msg << std::endl; \
+            std::stringstream errmsg{};                                                                \
+            errmsg << "[" #result "][" << __FILE__ << ":" << __LINE__ << "] " << msg " failure ("      \
+                   << status_msg << ")";                                                               \
+            throw std::runtime_error(errmsg.str());                                                    \
+        }                                                                                              \
+    }
+
+enum class iteration_multiplexing_mode_t
+{
+    DISABLED,
+    SIMPLE,
+    KERNEL,
+    LAUNCH
+};
+
+struct kernel_dispatch_info_t
+{
+    uint64_t           kernel_id;
+    uint64_t           queue_id;
+    rocprofiler_dim3_t workgroup_size;
+    rocprofiler_dim3_t grid_size;
+    uint32_t           LDS_memory_size;
+
+    bool operator<(const kernel_dispatch_info_t other) const
+    {
+        return std::tie(kernel_id,
+                        queue_id,
+                        workgroup_size.x,
+                        workgroup_size.y,
+                        workgroup_size.z,
+                        grid_size.x,
+                        grid_size.y,
+                        grid_size.z,
+                        LDS_memory_size) < std::tie(other.kernel_id,
+                                                    other.queue_id,
+                                                    other.workgroup_size.x,
+                                                    other.workgroup_size.y,
+                                                    other.workgroup_size.z,
+                                                    other.grid_size.x,
+                                                    other.grid_size.y,
+                                                    other.grid_size.z,
+                                                    other.LDS_memory_size);
+    }
+};
+
+struct iteration_multiplexing_dispatch_record_t
+{
+    std::size_t                                   config;
+    std::map<uint64_t, std::size_t>               kernel_config;
+    std::map<kernel_dispatch_info_t, std::size_t> dispatch_config;
+};
+
+struct counter_info_record_t
+{
+    uint64_t    dispatch_id;
+    uint64_t    agent_id;
+    uint64_t    kernel_id;
+    uint32_t    LDS_memory_size;
+    uint64_t    counter_id;
+    std::string counter_name;
+    double      counter_value;
+};
+
+struct tool_data_t
+{
+    std::mutex                                 mut{};
+    std::string                                output_filename{};
+    std::unordered_map<uint64_t, std::string>  counter_id_name_map{};
+    std::string                                requested_counters{};
+    std::string                                kernel_filter_include_regex{};
+    std::vector<std::pair<uint64_t, uint64_t>> kernel_filter_ranges{};
+    std::vector<counter_info_record_t>         counter_records;
+    std::set<uint64_t>                         target_kernel_ids{};
+    iteration_multiplexing_mode_t              iteration_multiplexing_mode{iteration_multiplexing_mode_t::DISABLED};
+};
+
+bool is_targetted_dispatch(const tool_data_t* tool, uint64_t kernel_id, uint64_t kernel_iteration);
 
 void dispatch_callback(rocprofiler_dispatch_counting_service_data_t dispatch_data,
                        rocprofiler_counter_config_id_t*             config,
